Cuda base and TODO for Stanislav

+++ Hittable/Sphere.h
@@ -9,14 +9,14 @@
 
 class Sphere : public Hittable {
 public:
-    Sphere(const Point3D& center, double radius)
+    Sphere(const Point3D& center, float radius)
         : mCenter(center), mRadius(radius) {}
 
-    bool hit(const Ray &ray, double tMin, double tMax, HitRecord &record) const override;
+    bool hit(const Ray &ray, float tMin, float tMax, HitRecord &record) const override;
 
 private:
     Point3D mCenter;
-    double mRadius;
+    float mRadius;
 };
 
 #endif //RAY_RACING_CLUB_SPHERE_H
